requirements.txt
asgiref==3.5.2
black==22.6.0
certifi==2022.6.15
click==8.1.3
cloudinary==1.29.0
colorama==0.4.5
Django==3.2.14
django-ckeditor==6.4.2
django-classy-tags==3.0.1
django-cms==3.10.1
django-crispy-forms==1.14.0
django-filter==22.1
django-formtools==2.3
django-jazzmin==2.5.0
django-js-asset==2.0.0
django-restframework==0.0.1
django-sekizai==4.0.0
django-treebeard==4.5.1
djangocms-admin-style==3.2.0
djangorestframework==3.13.1
mypy-extensions==0.4.3
pathspec==0.9.0
Pillow==9.2.0
platformdirs==2.5.2
pytz==2022.1
six==1.16.0
sqlparse==0.4.2
tomli==2.0.1
tzdata==2022.1
urllib3==1.26.11
